% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained-splines-greens.R
\name{constrSpline}
\alias{constrSpline}
\title{Constrained L2 spline.}
\usage{
constrSpline(t, y, gfun, mp = 100, constr = -1, sign = 1,
  lambda = 1e-04)
}
\arguments{
\item{t}{Numeric vector of measurement times.}

\item{y}{Numeric vector of values to be fitted.}

\item{gfun}{Choice of Greens function (choose from G1-G10).}

\item{mp}{Numeric value. Number of points to evaluate spline in.}

\item{constr}{Numeric value in (-1, 0, 1, 2).
Choice of constraint.
-1: No constraint.
0:  Positivity contraint.
1:  Monotonicity constraint.
2:  Convexity/concavity constraint.}

\item{sign}{Numeric value in (-1, 1).
1:  Increasing, convexity.
-1: Decreasing, concavity.}

\item{lambda}{Smoothness parameter.}
}
\description{
Function to compute a constrained L2 spline.
}
\examples{
t = generate.t(m = 25); y = generate.y(type = 2)(t) + generate.noise(t) + generate.noise(t)
plot(t, y, col = "blue")
S = constrSpline(t, y, G3, constr=1)
lines(attr(S, "t"), S)
t = generate.t(m = 25); y = -generate.y(type = 3)(t) + generate.noise(t) + generate.noise(t)
plot(t, y, col = "blue")
S = constrSpline(t, y, G3, constr=2, sign=-1)
lines(attr(S, "t"), S)
t = generate.t(m = 40); y = generate.y(type = 4)(t) + generate.noise(t)
plot(t, y, col = "blue")
S = constrSpline(t, y, G1, constr=0)
lines(t, S)
}

