% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions-laplace.R
\name{dLaplace}
\alias{dLaplace}
\title{Laplace density function.

The density function for the univariate Laplace distributions.
Can use both the parametrization with respect to kappa and mu.}
\usage{
dLaplace(x, mu = Inf, theta = 0, kappa = 1, sigma = 1)
}
\arguments{
\item{x}{Numeric value, where the function is evaluated.}

\item{mu}{Numeric value. If specified, this parametrization will be used.}

\item{theta}{Numeric value.}

\item{kappa}{Numeric value (positive).}

\item{sigma}{Numeric value (positive).}
}

