% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetafun-greens-functions.R
\name{G10}
\alias{G10}
\title{Greens function 10.}
\usage{
G10(s, t)
}
\arguments{
\item{s}{Numeric value.}

\item{t}{Numeric value.}
}
\description{
Function to compute the Greens function in two time points.
}

