% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetafun-greens-functions.R
\name{G3}
\alias{G3}
\title{Greens function 3.}
\usage{
G3(s, t)
}
\arguments{
\item{s}{Numeric value.}

\item{t}{Numeric value.}
}
\description{
Function to compute the Greens function in two time points.
}

