% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data-fun.R
\name{generate.data}
\alias{generate.data}
\title{Generate sample times.

Function to generate different kinds of sample times.}
\usage{
generate.data(m = 15, seed = sample(2500, 1), t.type = 1, y.type = 1,
  n.type = 1, mu = 0, theta = 0, kappa = Inf, sigma = 0.2,
  sigma2 = 1e-04, tfun = function(t) t/4)
}
\arguments{
\item{m}{Length of output vector. Default is 15.}

\item{seed}{If certain seed wanted.}

\item{t.type}{Numeric value (1,2).}

\item{y.type}{Numeric value (1,2,...,10).}

\item{n.type}{Numeric value (1,...,8).}

\item{mu}{Numeric value.}

\item{theta}{Numeric value.}

\item{kappa}{Numeric value.}

\item{sigma}{Numeric value.}

\item{sigma2}{Numeric value. Smaller SE for contaminated noise.}

\item{tfun}{For generating noise with dynamically changing parameters.}

\item{seed}{If certain seed wanted.}
}

