% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data-fun.R
\name{generate.noise}
\alias{generate.noise}
\title{Generating random noise.}
\usage{
generate.noise(t, type = 1, mu = 0, theta = 0, kappa = Inf,
  sigma = 0.2, sigma2 = 1e-04, seed = sample(2500, 1),
  tfun = function(t) t/4)
}
\arguments{
\item{t}{Numeric vector where to evaluate noise.}

\item{type}{Numeric value (1,...,8).}

\item{mu}{Numeric value.}

\item{theta}{Numeric value.}

\item{kappa}{Numeric value.}

\item{sigma}{Numeric value.}

\item{sigma2}{Numeric value. Smaller SE for contaminated noise.}

\item{seed}{If certain seed wanted.}

\item{tfun}{For generating noise with dynamically changing parameters.}
}
\description{
Function to generate different kinds of random noise.
}
\examples{
t = generate.t(m = 50)
w = generate.noise(t); plot(t, w , type="h", main = "Gaussian noise")
w = generate.noise(t, type = 2); plot(t, w , type="h", main = "Laplace noise")
w = generate.noise(t, type = 3); plot(t, w , type="h", main = "Contaminated noise")
w = generate.noise(t, type = 4); plot(t, w , type="h", main = "Uniform noise")
}

