% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data-fun.R
\name{generate.t}
\alias{generate.t}
\title{Generate sample times.

Function to generate different kinds of sample times.}
\usage{
generate.t(type = 1, m = 15, seed = sample(2500, 1))
}
\arguments{
\item{type}{Numeric value (1,2).}

\item{m}{Length of output vector.}

\item{seed}{If certain seed wanted.}
}
\examples{
generate.t()
}

