% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1-adapted-spline.R
\name{L1splineAdapted}
\alias{L1splineAdapted}
\title{Adapted L1 spline.}
\usage{
L1splineAdapted(x, t, y, gfun, sigma = sqrt(2), b = 0.5, lambda = 1e-04,
  niter = 200, stopiter = 1e-10)
}
\arguments{
\item{x}{Numeric value(s) to evaluate the spline in.}

\item{t}{Numeric vector of measurement times.}

\item{y}{Numeric vector of values to be fitted.}

\item{gfun}{Choice of Greens function (choose from G1-G10).}

\item{sigma}{Variance, per default set to srqt(2). Effect translated into lambda.}

\item{b}{Truncation window.}

\item{lambda}{Smoothness parameter.}

\item{niter}{Number of iterations.}

\item{stopiter}{Stopping criterion.}
}
\description{
Function to compute the adapted L1 spline, by use of the continuation method.
}

