% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1-AL-spline.R
\name{L1splineAL}
\alias{L1splineAL}
\title{Asymmetric L1 spline.}
\usage{
L1splineAL(x, t, y, gfun, kappa = NULL, lambda = 1e-04, CI = 0,
  niter = 200, stopiter = 1e-10, niterk = 5)
}
\arguments{
\item{x}{Numeric value(s) to evaluate the spline in.}

\item{t}{Numeric vector of measurement times.}

\item{y}{Numeric vector of values to be fitted.}

\item{gfun}{Choice of Greens function (choose from G1-G10).}

\item{kappa}{Asymmetry parameter. Per default kappa=NULL, which means that kappa will be estimated.}

\item{lambda}{Smoothness parameter.}

\item{CI}{Option to compute confidence intervals.
0: Nothing is computed.
2: sigmahat is obtain from kappa(hat) and thetahat, and CI is computed.}

\item{niter}{Number of iterations.}

\item{stopiter}{Stopping criterion.}

\item{niterk}{Number of iterations for the kappa estimation.}
}
\description{
Function to compute the asymmetric L1 spline, by use of the continuation method.
Assuming either kappa (asymmetry parameter) known, or unknown (to be estimated).
}

