% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1-split-bregman-iteration.R
\name{L1splineSB}
\alias{L1splineSB}
\title{Split Bregman L1 spline.}
\usage{
L1splineSB(x, t, y, gfun, lbratio = 0.001, beta = 1, niter = 200,
  stopiter = 1e-10)
}
\arguments{
\item{x}{Numeric value(s) to evaluate the spline in.}

\item{t}{Numeric vector of measurement times.}

\item{y}{Numeric vector of values to be fitted.}

\item{gfun}{Choice of Greens function (choose from G1-G10).}

\item{lbratio}{Numeric value. Ratio of lambda over beta.}

\item{beta}{Numeric value.}

\item{niter}{Number of iterations.}

\item{stopiter}{Stopping criterion.}
}
\description{
Function to compute the L1 spline, by use of the split Bregman method.
}

