% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1-sigma-unknown-spline.R
\name{L1splineSD}
\alias{L1splineSD}
\title{L1 spline (unknown dynamic variance).}
\usage{
L1splineSD(x, t, y, gfun, t0 = Inf, div = 10, lambda = 1e-04,
  niter = 200, stopiter = 1e-10, ...)
}
\arguments{
\item{x}{Numeric value(s) to evaluate the spline in.}

\item{t}{Numeric vector of measurement times.}

\item{y}{Numeric vector of values to be fitted.}

\item{gfun}{Choice of Greens function (choose from G1-G10).}

\item{t0}{Anchor points for changing kappa.
Per default set a division of [0,1] based on the number of observations.}

\item{div}{Integer determining where to put boundary knots.}

\item{lambda}{Smoothness parameter.}

\item{niter}{Number of iterations.}

\item{stopiter}{Stopping criterion.}

\item{sigma}{Variance, per default set to srqt(2). Effect translated into lambda.}
}
\description{
Function to compute an L1 spline, by use of the continuation method.
Fitting variance, allowing for dynamical change.
}

