% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2spline.R
\name{L2spline}
\alias{L2spline}
\title{Spline function.}
\usage{
L2spline(t, y, gfun, lambda = 0.001, x = Inf)
}
\arguments{
\item{t}{Numeric vector of observation times.}

\item{y}{Numeric vector to be fitted.}

\item{gfun}{Choice of Greens function (1-10).}

\item{lambda}{Smoothness parameter.}

\item{x}{Numeric vector. If specified, spline is computed in this set of points.}
}
\description{
Function to fit the L2 spline using some Greens function.
}
\examples{
t = generate.t(m = 25); y = generate.y()(t) + generate.noise(t, type = 2)
plot(t, y, col = "blue")
lines(t, thetafun(t, y, G9, lambda = 0.0001))
}

