% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2splineC.R
\name{L2splineC}
\alias{L2splineC}
\title{Constrained L2 spline.

Function to compute a constrained L2 spline.}
\usage{
L2splineC(t, y, gfun, mp = 100, constr = 0, sign = 1, lambda = 1e-04)
}
\arguments{
\item{t}{Numeric vector of measurement times.}

\item{y}{Numeric vector of values to be fitted.}

\item{gfun}{Choice of Greens function (choose from G1-G10).}

\item{mp}{Integer value. Number of points to evaluate spline in.}

\item{constr}{Integer value in (0, 1, 2).
Choice of constraint.
0:  Positivity contraint.
1:  Monotonicity constraint.
2:  Convexity/concavity constraint.}

\item{sign}{Numeric value in (-1, 1).
1:  Increasing, convexity.
-1: Decreasing, concavity.}

\item{lambda}{Smoothness parameter.}
}
\examples{
t = generate.t(m = 25); y = generate.y(type = 2)(t) + generate.noise(t) + generate.noise(t)
plot(t, y, col = "blue")
S = L2splineC(t, y, 3, constr=1)
lines(attr(S, "t"), S)
t = generate.t(m = 25); y = -generate.y(type = 3)(t) + generate.noise(t) + generate.noise(t)
plot(t, y, col = "blue")
S = L2splineC(t, y, 3, constr=2, sign=-1)
lines(attr(S, "t"), S)
t = generate.t(m = 40); y = generate.y(type = 4)(t) + generate.noise(t)
plot(t, y, col = "blue")
S = L2splineC(t, y, 1, constr=0)
lines(t, S)
}

