% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{MSEfun}
\alias{MSEfun}
\title{Mean squared error.}
\usage{
MSEfun(y, yhat)
}
\arguments{
\item{y}{Numeric vector.}

\item{yhat}{Numeric vector.}
}
\description{
Function that computes the mean squared error of two vectors.
}

