% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outerG}
\alias{outerG}
\title{Outer function for gfun.}
\usage{
outerG(t, tp, gfun, inv)
}
\arguments{
\item{t}{Numeric vector of observation times.}

\item{tp}{Numeric vector evaluation times.}

\item{gfun}{Integer value 1-10 (choice of Greens function).}

\item{inv}{Logical (want inverse matrix or not)}
}
\value{
Numeric vector of fitted values.
}
\description{
Function to compute the matrix of Green's function evaluation for any two vectors t and tp.
}
\author{
Helene Charlotte Rytgaard
}

