% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative-distribution-laplace.R, R/distribution-functions-laplace.R
\name{qLaplace}
\alias{qLaplace}
\title{Laplace quantile function.

The quantile function for the univariate Laplace distributions.
Obtained from numerically inverting pLaplace.
The parametrization with kappa is used}
\usage{
qLaplace(q, mu = Inf, theta = 0, kappa = 1, sigma = 1)

qLaplace(q, mu = Inf, theta = 0, kappa = 1, sigma = 1)
}
\arguments{
\item{q}{Numeric value, where the function is evaluated, between 0 and 1.}

\item{mu}{Numeric value. If specified, this parametrization will be used.}

\item{theta}{Numeric value.}

\item{kappa}{Numeric value (positive).}

\item{sigma}{Numeric value (positive).}

\item{q}{Numeric value, where the function is evaluated, between 0 and 1.}

\item{theta}{Numeric value.}

\item{kappa}{Numeric value (positive).}

\item{sigma}{Numeric value (positive).}
}

