% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative-distribution-laplace.R, R/distribution-functions-laplace.R
\name{rLaplace}
\alias{rLaplace}
\title{Laplace simulation function.

Function that simulates from the univariate Laplace distributions.
Two parametrizations can be used.}
\usage{
rLaplace(m, mu = Inf, theta = 0, kappa = 1, sigma = 1)

rLaplace(m, mu = Inf, theta = 0, kappa = 1, sigma = 1)
}
\arguments{
\item{m}{Numeric value, length of t vector.}

\item{mu}{Numeric value.}

\item{theta}{Numeric value.}

\item{kappa}{Numeric value.
If not specified: Use one parametrization.
If specified: Use other parametrization.}

\item{sigma}{Numeric value (positive).}

\item{m}{Numeric value, length of t vector.}

\item{mu}{Numeric value. If specified, this parametrization will be used.}

\item{theta}{Numeric value.}

\item{kappa}{Numeric value (positive).}
}

