% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace-multivariate-simulation-function.R
\name{rMultiLaplace}
\alias{rMultiLaplace}
\title{Laplace simulation function.

Function that simulates from the univariate Laplace distributions.
Two parametrizations can be used.}
\usage{
rMultiLaplace(n = 1, theta = c(0, 0), gamma = c(0, 0), Sigma = diag(2),
  sigma = 0.2, tau = 1, seed = sample(2500, 1))
}
\arguments{
\item{n}{Numeric value, number of simulations.}

\item{theta}{Numeric vector of position parameters of length d.}

\item{gamma}{Numeric vector of skewness parameters of length d.}

\item{Sigma}{Numeric positive definite matrix of size d times d.}

\item{sigma}{Numeric value (positive).}

\item{tau}{Numeric value (positive).}

\item{seed}{Integer value to be specified if want certain seed.}
}

