% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-multivar-laplace.R
\name{rMultLaplace}
\alias{rMultLaplace}
\title{thetaltivariate Laplace sithetalation function.}
\usage{
rMultLaplace(m, theta, gamma, Sigma, sigma = 0.2, tau = 1)
}
\arguments{
\item{m}{Numeric value, length of t vector.}

\item{theta}{Numeric vector with length m.}

\item{Sigma}{Matrix (m times m).}

\item{sigma}{Numeric value (positive).}

\item{tau}{Numeric value (positive).}

\item{theta}{Numeric vector with length m.}

\item{kappa}{Numeric value with length m.}
}
\description{
Function that sithetalates from the thetaltivariate Laplace distributions.
}

