% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{theme.p}
\alias{theme.p}
\title{ggplot2 theme.}
\usage{
theme.p(p, ls = 20, ts = 15, as = 15)
}
\arguments{
\item{p}{The ggplot2 object.}

\item{ls}{Legend size (default = 20).}

\item{ts}{Title size (default = 15).}

\item{as}{Axis text size (default = 15).}
}
\description{
User-friendly ggplot2 theme.
}

