% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetafun-greens-functions.R
\name{thetafun}
\alias{thetafun}
\title{Spline function.}
\usage{
thetafun(t, y, gfun, x = Inf, lambda = 0.001)
}
\arguments{
\item{t}{Numeric vector of evaluation times.}

\item{y}{Numeric vector to be fitted.}

\item{gfun}{Choice of Greens function.}

\item{x}{If numeric value then the L2 spline will be evaluated here.
Otherwise it will be evaluated in t.}

\item{lambda}{Smoothness parameter.}
}
\description{
Function to fit the L2 spline using some Greens function.
}
\examples{
t = generate.t(m = 25); y = generate.y()(t) + generate.noise(t, type = 2)
plot(t, y, col = "blue")
lines(t, thetafun(t, y, G9, lambda = 0.0001))
}

