% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{thetafunG}
\alias{thetafunG}
\title{General solution, Green's spline.}
\usage{
thetafunG(t, y, gfun, lambda)
}
\arguments{
\item{t}{Numeric vector of observation times.}

\item{y}{Numeric vector of observations.}

\item{gfun}{Integer value 1-10 (choice of Greens function).}

\item{lambda}{Numeric (smoothness parameter).}
}
\value{
Numeric vector of fitted values.
}
\description{
Function that takes the observations and computes the general spline solution. Must be combined with outerG() to obtain solution in some set of points.
}
\author{
Helene Charlotte Rytgaard
}

