% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative-distribution-laplace.R
\name{dLaplace}
\alias{dLaplace}
\title{Laplace density function.

The density function for the univariate Laplace distributions.
Can use both the parametrization with respect to kappa and mu.}
\usage{
dLaplace(x, mu = 1, sigma = 0.5, kappa = Inf, theta = 0.5)
}
\arguments{
\item{x}{Numeric value, where the function is evaluated.}

\item{mu}{Numeric value.}

\item{sigma}{Numeric value (positive).}

\item{kappa}{Numeric value (positive).
if not specified, parametrization with mu is used.
if specified (numeric value), parametrization with kappa used.}

\item{theta}{Numeric value.}
}

