% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data-fun.R
\name{generate.y}
\alias{generate.y}
\title{Generating observations (underlying functions).}
\usage{
generate.y(type = 1, seed = sample(2500, 1))
}
\arguments{
\item{type}{Numeric value (1,2,...,10).}

\item{seed}{If certain seed wanted.}
}
\description{
Function to generate different kinds of underlying functions.
}
\examples{
y = generate.y(); plot(y)
y = generate.y(type = 2); plot(y)
y = generate.y(type = 3); plot(y)
y = generate.y(type = 4); plot(y)
}

