% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1-continuation-method.R
\name{L1spline}
\alias{L1spline}
\title{L1 spline.}
\usage{
L1spline(x, t, y, gfun, sigma = sqrt(2), lambda = 1e-04, niter = 200,
  stopiter = 1e-10, constr = -Inf, sign = 1, mp = 100)
}
\arguments{
\item{x}{Numeric value(s) to evaluate the spline in.}

\item{t}{Numeric vector of measurement times.}

\item{y}{Numeric vector of values to be fitted.}

\item{gfun}{Choice of Greens function (choose from G1-G10).}

\item{sigma}{Variance, per default set to srqt(2). Effect translated into lambda.}

\item{lambda}{Smoothness parameter.}

\item{niter}{Number of iterations.}

\item{stopiter}{Stopping criterion.}

\item{constr}{Numeric value in (-1, 0, 1, 2).
Choice of constraint.
-1: No constraint.
0:  Positivity contraint.
1:  Monotonicity constraint.
2:  Convexity/concavity constraint.}

\item{sign}{Numeric value in (-1, 1).
1:  Increasing, convexity.
-1: Decreasing, concavity.}

\item{mp}{Numeric value. Number of points to evaluate spline in.}
}
\description{
Function to compute the L1 spline, by use of the continuation method.
}
\examples{
t = generate.t(m = 25); y = generate.y()(t) + generate.noise(t, type = 2)
plot(t, y, col = "blue")
lines(t, L1spline(t, t, y, G9, lambda = 0.0001))
}

