% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{L2spline}
\alias{L2spline}
\title{something something}
\usage{
L2spline(t, y, gfun, lambda)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A matrix of same size as x.
}
\author{
Helene Charlotte Rytgaard
}

