% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative-distribution-laplace.R
\name{pLaplace}
\alias{pLaplace}
\title{Laplace cumulative distribution function.}
\usage{
pLaplace(x, theta = 0, kappa = 1, sigma = 1)
}
\arguments{
\item{x}{Numeric value, where the function is evaluated.}

\item{theta}{Numeric value.}

\item{kappa}{Numeric value (positive).}

\item{sigma}{Numeric value (positive).}
}
\description{
The cumulative distribution function for the univariate Laplace distributions.
The parametrization with kappa is used
}

