% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative-distribution-laplace.R
\name{qLaplace}
\alias{qLaplace}
\title{Laplace quantile function.}
\usage{
qLaplace(q, theta = 0, kappa = 1, sigma = 1)
}
\arguments{
\item{q}{Numeric value, where the function is evaluated, between 0 and 1.}

\item{theta}{Numeric value.}

\item{kappa}{Numeric value (positive).}

\item{sigma}{Numeric value (positive).}
}
\description{
The quantile function for the univariate Laplace distributions.
Obtained from numerically inverting pLaplace.
The parametrization with kappa is used
}

