% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace-sim.R
\name{rLaplace}
\alias{rLaplace}
\title{Laplace simulation function.}
\usage{
rLaplace(m, mu = 0, theta = 0, kappa = Inf, sigma = 1)
}
\arguments{
\item{m}{Numeric value, length of t vector.}

\item{mu}{Numeric value.}

\item{theta}{Numeric value.}

\item{kappa}{Numeric value.
If not specified: Use one parametrization.
If specified: Use other parametrization.}
}
\description{
Function that simulates from the univariate Laplace distributions.
Two parametrizations can be used.
}

