% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{thetafun}
\alias{thetafun}
\title{Spline function.}
\usage{
thetafun(t, y, gfun, lambda)
}
\arguments{
\item{t}{Numeric vector of evaluation times.}

\item{y}{Numeric vector to be fitted.}

\item{gfun}{Integer value 1-10 (choice of Greens function).}

\item{lambda}{Numeric (smoothness parameter).}
}
\value{
Numeric vector of fitted values.
}
\description{
Function to fit the L2 spline using some Greens function.
}
\examples{
t = generate.t(m = 25); y = generate.y()(t) + generate.noise(t, type = 2)
plot(t, y, col = "blue")
lines(t, thetafun(t, y, 9, lambda = 0.0001))
}
\author{
Helene Charlotte Rytgaard
}

